LinkedIn API Added to Feedcrunch

+++ feedcrunch_rssadmin/static/js/admin-social-linking.js
@@ -2,19 +2,37 @@ $(document).ready(function() {
 
 	var interval;
 
-	function checkTwitterLinkStatus() {
+	function checkSocialNetworkStatus(social_network) {
+        var social_networks_dict = {
+            twitter: "#twitter-btn-div",
+            facebook: "#facebook-btn-div",
+            linkedin: "#linkedin-btn-div",
+            gplus: "#gplus-btn-div",
+        };
+
+        if (Object.keys(social_networks_dict).indexOf(social_network) == -1){
+            console.log("This social network is not supported: " + social_network);
+            return False;
+        }
+
 		$.ajax({
 			type: 'GET',
-			url: '/api/1.0/authenticated/get/user/social-networks/twitter/status/',
+			url: '/api/1.0/authenticated/get/user/social-networks/'+social_network+'/status/',
 			dataType: 'json',
 			success: function(data){
 				if (data.success) {
 					if (data.status){
-						$("#twitter-btn-div").html('<button class="waves-effect waves-light btn red twitter-btn" id="twitter-btn-unlink">Unlink My Twitter Account</button>');
+                        $(social_networks_dict[social_network]).html(
+                            `<button class="waves-effect waves-light btn red socialconnect-btn" id="`+social_network+`-btn-unlink">
+                                Unlink My `+social_network[0].toUpperCase() + social_network.slice(1)+` Account
+                            </button>`
+                        );
 						setUnlinkClickEvent();
 					}
 					else
-						interval = setTimeout(checkTwitterLinkStatus, 1500);
+            			interval = setTimeout(function() {
+                            checkSocialNetworkStatus(social_network);
+                        }, 1500);
 				}
 				else{
 					swal({
@@ -30,74 +48,35 @@ $(document).ready(function() {
 		});
 	}
 
-    function UnLinkTwitter() {
+    function UnLinkSocialNetwork(social_network) {
         var csrftoken = Cookies.get('csrftoken');
+
+        var social_networks_dict = {
+            twitter: "#twitter-btn-div",
+            facebook: "#facebook-btn-div",
+            linkedin: "#linkedin-btn-div",
+            gplus: "#gplus-btn-div",
+        };
+
+        if (Object.keys(social_networks_dict).indexOf(social_network) == -1){
+            console.log("This social network is not supported: " + social_network);
+            return False;
+        }
+
         $.ajax({
             type: 'DELETE',
-            url: '/api/1.0/authenticated/delete/user/social-networks/twitter/',
+            url: '/api/1.0/authenticated/delete/user/social-networks/'+social_network+'/',
             dataType: 'json',
             beforeSend: function(xhr) {
                 xhr.setRequestHeader("X-CSRFToken", csrftoken);
             },
             success: function(data){
                 if (data.success) {
-                    $("#twitter-btn-div").html('<a href="' + data.auth_url + '" target="_target" class="waves-effect waves-light btn blue" id="twitter-btn-link">Link My Twitter Account</a>');
-                    setLinkClickEvent();
-                }
-                else{
-                    swal({
-                        title: "Something went wrong!",
-                        text: data.error,
-                        type: "error",
-                        confirmButtonColor: "#DD6B55",
-                        confirmButtonText: "I'll Fix it!",
-                        closeOnConfirm: true
-                    });
-                }
-            }
-        });
-    }
-
-    function checkFacebookLinkStatus() {
-		$.ajax({
-			type: 'GET',
-			url: '/api/1.0/authenticated/get/user/social-networks/facebook/status/',
-			dataType: 'json',
-			success: function(data){
-				if (data.success) {
-					if (data.status){
-						$("#facebook-btn-div").html('<button class="waves-effect waves-light btn red twitter-btn" id="facebook-btn-unlink">Unlink My Facebook Account</button>');
-						setUnlinkClickEvent();
-					}
-					else
-						interval = setTimeout(checkFacebookLinkStatus, 1500);
-				}
-				else{
-					swal({
-						title: "Something went wrong!",
-						text: data.error,
-						type: "error",
-						confirmButtonColor: "#DD6B55",
-						confirmButtonText: "I'll Fix it!",
-						closeOnConfirm: true
-					});
-				}
-			}
-		});
-	}
-
-    function UnLinkFacebook() {
-        var csrftoken = Cookies.get('csrftoken');
-        $.ajax({
-            type: 'DELETE',
-            url: '/api/1.0/authenticated/delete/user/social-networks/facebook/',
-            dataType: 'json',
-            beforeSend: function(xhr) {
-                xhr.setRequestHeader("X-CSRFToken", csrftoken);
-            },
-            success: function(data){
-                if (data.success) {
-                    $("#facebook-btn-div").html('<a href="' + data.auth_url + '"  target="_target" class="waves-effect waves-light btn blue" id="facebook-btn-link">Link My Facebook Account</a>');
+                    $(social_networks_dict[social_network]).html(
+                        `<a href="`+ data.auth_url +`"  target="_target" class="waves-effect waves-light btn blue socialconnect-btn" id="`+social_network+`-btn-link">
+                            Link My `+social_network[0].toUpperCase() + social_network.slice(1)+` Account
+                        </a>`
+                    );
                     setLinkClickEvent();
                 }
                 else{
@@ -117,23 +96,60 @@ $(document).ready(function() {
     /* ========================== Event Callbacks ***************************** */
 
 	function setLinkClickEvent() {
+        $("#twitter-btn-link").unbind("click");
 		$("#twitter-btn-link").click(function() {
-			interval = setTimeout(checkTwitterLinkStatus, 1500);
+			interval = setTimeout(function() {
+                checkSocialNetworkStatus("twitter");
+            }, 1500);
 		});
+
+        $("#facebook-btn-link").unbind("click");
         $("#facebook-btn-link").click(function() {
-			interval = setTimeout(checkFacebookLinkStatus, 1500);
+			interval = setTimeout(function() {
+                checkSocialNetworkStatus("facebook");
+            }, 1500);
+		});
+
+        $("#linkedin-btn-link").unbind("click");
+        $("#linkedin-btn-link").click(function() {
+			interval = setTimeout(function() {
+                checkSocialNetworkStatus("linkedin");
+            }, 1500);
+		});
+
+        $("#gplus-btn-link").unbind("click");
+        $("#gplus-btn-link").click(function() {
+			interval = setTimeout(function() {
+                checkSocialNetworkStatus("gplus");
+            }, 1500);
 		});
 	}
 
 	function setUnlinkClickEvent() {
+        $("#twitter-btn-unlink").unbind("click");
 		$("#twitter-btn-unlink").click(function(event) {
-            UnLinkTwitter();
+            UnLinkSocialNetwork("twitter");
             event.preventDefault();
             return false;
         });
 
+        $("#facebook-btn-unlink").unbind("click");
         $("#facebook-btn-unlink").click(function(event) {
-            UnLinkFacebook();
+            UnLinkSocialNetwork("facebook");
+            event.preventDefault();
+            return false;
+        });
+
+        $("#linkedin-btn-unlink").unbind("click");
+        $("#linkedin-btn-unlink").click(function(event) {
+            UnLinkSocialNetwork("linkedin");
+            event.preventDefault();
+            return false;
+        });
+
+        $("#gplus-btn-unlink").unbind("click");
+        $("#gplus-btn-unlink").click(function(event) {
+            UnLinkSocialNetwork("gplus");
             event.preventDefault();
             return false;
         });
